+++ libsmartstack/report.h
@@ -5,13 +5,14 @@
 #ifndef SMARTSTACK_LIBSMARTSTACK_REPORT_H_
 #define SMARTSTACK_LIBSMARTSTACK_REPORT_H_
 
-#include "function.h"
 #include <string>
 #include <vector>
 
+#include "function.h"
+
 namespace SmartStack {
 class Report {
-public:
+ public:
   enum SortOrder { Ascending, Descending };
   enum SortType { Time, MeanTime, TotalTime, MeanTotalTime, Calls };
   enum TimeUnits { Microseconds, Milliseconds, Seconds, Minutes, Hours };
@@ -33,33 +34,31 @@ public:
   void saveCsvTimingReport(const std::vector<Function *> &functions,
                            const std::string &filename) const;
 
-private:
-  [[nodiscard]] std::vector<Function *>
-  sortFunctions(const std::vector<Function *> &functions) const;
+ private:
+  [[nodiscard]] std::vector<Function *> sortFunctions(
+      const std::vector<Function *> &functions) const;
 
   void getSortCodes(std::string &calls, std::string &duration,
                     std::string &meanDuration, std::string &totalDuration,
                     std::string &meanTotalDuration, const SortType &st,
                     const SortOrder &so) const;
 
-  [[nodiscard]] std::vector<std::string>
-  generateTableTimingReport(const std::vector<Function *> &functions,
-                            const SortType &st, const SortOrder &so) const;
+  [[nodiscard]] std::vector<std::string> generateTableTimingReport(
+      const std::vector<Function *> &functions, const SortType &st,
+      const SortOrder &so) const;
 
   [[nodiscard]] std::string unitsString(const Report::TimeUnits &units,
                                         bool trim) const;
 
-  [[nodiscard]] std::string
-  getFunctionReportLine(size_t i, Function *f, size_t function_name_max_length,
-                        const Report::TimeUnits &units,
-                        const Report::OutputFormat &format) const;
+  [[nodiscard]] std::string getFunctionReportLine(
+      size_t i, Function *f, size_t function_name_max_length,
+      const Report::TimeUnits &units, const Report::OutputFormat &format) const;
 
-  [[nodiscard]] static double convertTimeUnitsDouble(size_t time,
+  [[nodiscard]] static double convertTimeUnitsDouble(long long time,
                                                      double multiplier);
 
-  [[nodiscard]] static size_t
-  maxNumFunctionChars(const std::vector<Function *> &functions,
-                      size_t lowerLimit);
+  [[nodiscard]] static size_t maxNumFunctionChars(
+      const std::vector<Function *> &functions, size_t lowerLimit);
 
   [[nodiscard]] static std::string formatStringChar(size_t n);
 
@@ -68,6 +67,6 @@ private:
   SortOrder m_sortOrder;
   SortType m_sortType;
 };
-} // namespace SmartStack
+}  // namespace SmartStack
 
-#endif // SMARTSTACK_LIBSMARTSTACK_REPORT_H_
+#endif  // SMARTSTACK_LIBSMARTSTACK_REPORT_H_
