+++ libsmartstack/timer.cpp
@@ -42,8 +42,8 @@ void Timer::stopClock() {
     this->m_endLocal = std::chrono::high_resolution_clock::now();
     this->m_endGlobal = this->m_endLocal;
     this->m_lastElapsed = std::chrono::duration_cast<std::chrono::microseconds>(
-                              this->m_endLocal - this->m_startLocal)
-                              .count();
+        this->m_endLocal - this->m_startLocal)
+        .count();
     this->m_totalElapsed += this->m_lastElapsed;
     this->m_totalGlobalElapsed +=
         std::chrono::duration_cast<std::chrono::microseconds>(
@@ -59,8 +59,8 @@ void Timer::pause() {
   if (this->m_running) {
     this->m_endLocal = std::chrono::high_resolution_clock::now();
     this->m_lastElapsed = std::chrono::duration_cast<std::chrono::microseconds>(
-                              this->m_endLocal - this->m_startLocal)
-                              .count();
+        this->m_endLocal - this->m_startLocal)
+        .count();
     this->m_totalElapsed += this->m_lastElapsed;
   }
 }
@@ -76,21 +76,21 @@ void Timer::restart() {
  * @brief Timer::elapsed Returns the elapsed time in microseconds
  * @return Elapsed time in microseconds
  */
-size_t Timer::elapsed() const { return this->m_totalElapsed; }
+long long Timer::elapsed() const { return this->m_totalElapsed; }
 
 /**
  * @brief Timer::globalElapsed Returns the elapsed time in microseconds
  * @return Elapsed time in microseconds
  */
-size_t Timer::globalElapsed() const { return this->m_totalGlobalElapsed; }
+long long Timer::globalElapsed() const { return this->m_totalGlobalElapsed; }
 
 /**
  * @brief Timer::startTime Returns the start time in milliseconds
  * @return Start time in milliseconds
  */
-size_t Timer::startTime() const {
+long long Timer::startTime() const {
   return std::chrono::time_point_cast<std::chrono::milliseconds>(
-             this->m_startLocal)
+      this->m_startLocal)
       .time_since_epoch()
       .count();
 }
@@ -99,9 +99,9 @@ size_t Timer::startTime() const {
  * @brief Timer::endTime Returns the end time in milliseconds
  * @return End time in milliseconds
  */
-size_t Timer::endTime() const {
+long long Timer::endTime() const {
   return std::chrono::time_point_cast<std::chrono::milliseconds>(
-             this->m_endLocal)
+      this->m_endLocal)
       .time_since_epoch()
       .count();
 }
@@ -116,4 +116,4 @@ bool Timer::running() const { return this->m_running; }
  * @brief Timer::lastElapsed Returns the last elapsed time in microseconds
  * @return Last elapsed time in microseconds
  */
-size_t Timer::lastElapsed() const { return this->m_lastElapsed; }
+long long Timer::lastElapsed() const { return this->m_lastElapsed; }
